Moved tweet-related class to their own package.
