+++ src/main/java/io/anserini/nrts/TweetSearcher.java
@@ -19,12 +19,9 @@ import org.apache.lucene.store.MMapDirectory;
 
 import org.eclipse.jetty.server.Server;
 import org.eclipse.jetty.servlet.ServletContextHandler;
-import org.eclipse.jetty.servlet.ServletHandler;
 import org.eclipse.jetty.server.handler.ResourceHandler;
 import org.eclipse.jetty.server.handler.HandlerList;
-import io.anserini.nrts.TweetServlet;
-import io.anserini.index.twitter.TweetAnalyzer;
-import io.anserini.nrts.TweetSearcherAPI;
+import io.anserini.analysis.TweetAnalyzer;
 import org.eclipse.jetty.servlet.ServletHolder;
 import org.glassfish.jersey.servlet.ServletContainer;
